experimenting with animation blending (eg: fast swap to another animation)

+++ source/StateTransition.h
@@ -9,19 +9,11 @@
 
 	// TODO: std::sort the stateTransitions vector according to anyState bool AFTER all transitions have been loaded, then update transitionHash (in eAnimController)
 
-	// TODO: transitions can EITHER occur at any point, or wait until the state normalizedTime == ExitTime [==1.0f usually]),
-	// and only check the transition parameter values AFTER normalizedTime >= ExitTime (if any)
-
-	// TODO: eTransitions always start the new state from the beginning of its animation (regardless of "Exit Time") (unless I also add a transition offset for where to start the new normalizedTime),
-	// while eBlendStates start the selected animation at the same normalized time, right away
+	// TODO: eBlendStates should start the selected animation at the same normalized time
 
 	// TODO: create a base eStateNode class, and derive eAnimationState and eBlendState
 	// eAnimationState will have only one eAnimation to Update
-	// eBlendState will have a vector of eAnimations to select from based on 1 or 2 (not more) float_type parameters (only floats, not ints or bools)
-	// from its eAnimationController owner (as float *) (NOTE: at least one float must be selected, if not in load-file, then default to first floatParameter)
-	// AND the selection will be based on lowest squaredDistance from controllerParam(s) position to blendNode param(s) position
-	// AND the blendNode params will be 1 or 2 floats PER eAnimation in the eBlendState (directly proportional to controllerParams being tracked)
-	// and associated fixed-values (for getting the distance to)
+
 */
 
 //**************************************
@@ -54,7 +46,7 @@ private:
 private:
 
 	std::string				name;
-	size_t					nameHash;
+	int						nameHash;
 	float					exitTime;			// currentState::normalizedTime to start checking conditions (if <= 0.0f, there MUST be at least one transition param)
 	float					offset;				// the normalizedTime to start playing at in toState
 	bool					anyState;			// all conditions checked regardless of eAnimationController::currentState (DEBUG: ie: ignores fromState)
